+++ boilerplates/boilerplate-vue-ts/pages/_default/_default.page.server.ts
@@ -7,12 +7,11 @@ import logoUrl from './logo.svg'
 export { render }
 export { passToClient }
 
-// We use `contextProps.pageProps` to hold the props of the root component.
-// We pass `contextProps.pageProps` to the browser for `hydrate()` in `_default.page.client.ts`.
-const passToClient = ['pageProps']
+// See https://github.com/brillout/vite-plugin-ssr#data-fetching
+const passToClient = ["pageProps", "routeParams"];
 
 async function render({ Page, contextProps }: { Page: VueComponent; contextProps: ContextProps }) {
-  const app = createApp(Page, contextProps.pageProps)
+  const app = createApp(Page, contextProps)
   const appHtml = await renderToString(app)
   const title = 'My Vite SSR app'
   return html`<!DOCTYPE html>