+++ boilerplates/boilerplate-vue/pages/_default/app.js
@@ -3,7 +3,7 @@ import PageLayout from './PageLayout.vue'
 
 export { createApp }
 
-function createApp(Page, pageProps = {}) {
+function createApp(Page, contextProps) {
   const PageWithLayout = {
     render() {
       return h(
@@ -11,12 +11,19 @@ function createApp(Page, pageProps = {}) {
         {},
         {
           default() {
-            return h(Page, pageProps)
+            return h(Page, contextProps.pageProps || {})
           }
         }
       )
     }
   }
+
   const app = createSSRApp(PageWithLayout)
+
+  // We make `contextProps.routeParams` available in all components as
+  // `this.$routeParams` (e.g. `this.$routeParams.movieId` for a Route
+  // String `/movie/:movieId`).
+  app.config.globalProperties.$routeParams = contextProps.routeParams
+
   return app
 }