+++ src/client/router/getContextProps.client.ts
@@ -1,10 +1,9 @@
 import { navigationState } from '../navigationState.client'
-import { assert, assertUsage, getFileUrl, hasProp } from '../../utils'
+import { addUrlToContextProps, assert, assertUsage, getFileUrl, hasProp } from '../../utils'
 import { parse } from '@brillout/json-s'
 import { getPageInfo as getOriginalPageInfo } from '../getPage.client'
 
 export { getContextProps }
-export { retrieveContextProps }
 
 async function getContextProps(
   url: string,
@@ -30,5 +29,6 @@ async function retrieveContextProps(url: string): Promise<Record<string, unknown
   assert(hasProp(responseObject, 'contextProps'))
   const { contextProps } = responseObject
   assert(contextProps.constructor === Object)
+  addUrlToContextProps(contextProps, url)
   return contextProps
 }